ipythonblocks: Convert Block and Pixel to subclasses of _Item
This separates all of the indexing boilerplate from the actual block
formatting code, which makes both more resilient and easier to
understand.

I use ._index instead of using ._row and ._col, to make it easier to
scale to arbirary dimensional grids.  The differences in row/column
ordering between the two classes are now encoded in a new class-level
._row_major attribute.

I also added .index_string() to avoid duplicating a stringification
algorithm that scales to an arbitrary shape.